Move incentive saving logic to useIncentive
