+++ app/javascript/components/ResearcherApp/IncentiveForm.tsx
@@ -1,79 +1,8 @@
-import React, { useCallback, useReducer, useState } from "react";
-import { updateIncentive } from "@api/endpoints";
-
-type State = { incentive: Incentive } & (
-  | { status: "saved"; message: "Successfully updated!" }
-  | { status: "saving" }
-  | { status: "error"; error: Error; message: "An error occurred" }
-  | { status: "waiting" }
-  | { status: "changing" }
-);
-
-type Action =
-  | { type: "saved"; payload: Incentive }
-  | { type: "saving" }
-  | { type: "error"; payload: Error }
-  | {
-      type: "change";
-      payload: { index: number; code: Incentive["codes"][number] };
-    };
-
-const reducer = (state: State, action: Action): State => {
-  switch (action.type) {
-    case "error":
-      return {
-        status: "error",
-        error: action.payload,
-        incentive: state.incentive,
-        message: "An error occurred",
-      };
-    case "saved":
-      return {
-        status: "saved",
-        incentive: action.payload,
-        message: "Successfully updated!",
-      };
-    case "saving":
-      return { status: "saving", incentive: state.incentive };
-    case "change":
-      return {
-        status: "changing",
-        incentive: {
-          ...state.incentive,
-          codes: [
-            ...state.incentive.codes.slice(0, action.payload.index),
-            action.payload.code,
-            ...state.incentive.codes.slice(action.payload.index + 1),
-          ],
-        },
-      };
-  }
-};
+import React from "react";
+import { useIncentive } from "../../hooks";
 
 export const IncentiveForm = ({ incentive }: { incentive: Incentive }) => {
-  const [state, dispatch] = useReducer(reducer, {
-    incentive,
-    status: "waiting",
-  });
-
-  const save = async () => {
-    dispatch({ type: "saving" });
-
-    try {
-      const payload = await updateIncentive(state.incentive);
-
-      dispatch({ type: "saved", payload });
-    } catch (error) {
-      dispatch({ type: "error", payload: error });
-    }
-  };
-
-  const addCode = () => {
-    dispatch({
-      type: "change",
-      payload: { index: state.incentive.codes.length, code: "" },
-    });
-  };
+  const [state, { save, addCode, handleInputChange }] = useIncentive(incentive);
 
   return (
     <div>
@@ -85,12 +14,7 @@ export const IncentiveForm = ({ incentive }: { incentive: Incentive }) => {
             type="text"
             name="incentive_code"
             value={code}
-            onChange={(e) =>
-              dispatch({
-                type: "change",
-                payload: { index, code: e.currentTarget.value },
-              })
-            }
+            onChange={handleInputChange(index)}
           />
         </div>
       ))}
