+++ app/javascript/hooks/index.ts
@@ -1,1 +1,2 @@
+export * from "./use-incentive";
 export * from "./use-incentives";
